Improve division of responsibilities r.e. emails
